original-backend-boilerplate/backend/requirements.txt
fastapi>=0.109.0,<1
slowapi>=0.1.9,<1
google-genai>=1.0.0,<2
httpx>=0.26.0,<1
uvicorn[standard]>=0.27.0,<1
python-multipart>=0.0.6,<1
python-dotenv>=1.0.0,<2
supabase>=2.3.0,<3
twilio>=9.0.0,<10
pydantic>=2.5.0,<3
